fix bug in shoppingDao

+++ src/main/java/org/yearup/controllers/CategoriesController.java
@@ -1,8 +1,10 @@
 package org.yearup.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.yearup.data.CategoryDao;
 import org.yearup.data.ProductDao;
 import org.yearup.models.Category;
@@ -42,14 +44,28 @@ public class CategoriesController
     }
 
     // add the appropriate annotation for a get action
+//    @GetMapping("/{id}")
+//    public Category getById(@PathVariable int id)
+//    { return categoryDao.getById(id); }
+        // get the category by id
+
     @GetMapping("/{id}")
     public Category getById(@PathVariable int id)
-    { return categoryDao.getById(id); }
-        // get the category by id
+    {
+        Category category = categoryDao.getById(id);
+
+        if (category == null)
+        {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+        return category;
+    }
+
+
 
     // the url to return all products in category 1 would look like this
     // https://localhost:8080/categories/1/products
-    @GetMapping("{categoryId}/products")
+    @GetMapping("/{categoryId}/products")
     public List<Product> getProductsById(@PathVariable int categoryId)
     {
         // get a list of product by categoryId
@@ -58,11 +74,12 @@ public class CategoriesController
 
     // add annotation to call this method for a POST action
     // add annotation to ensure that only an ADMIN can call this function
+
     @PostMapping
     @PreAuthorize("hasRole('ADMIN')")
+    @ResponseStatus(HttpStatus.CREATED)
     public Category addCategory(@RequestBody Category category)
     {
-        // insert the category
         return categoryDao.create(category);
     }
 
@@ -72,6 +89,10 @@ public class CategoriesController
     @PreAuthorize("hasRole('ADMIN')")
     public void updateCategory(@PathVariable int id, @RequestBody Category category)
     { // update the category by id
+//        categoryDao.update(id, category);
+        if (categoryDao.getById(id) == null)
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
         categoryDao.update(id, category);
     }
 
@@ -79,10 +100,15 @@ public class CategoriesController
     // add annotation to ensure that only an ADMIN can call this function
     // DELETE http://localhost:8080/categories/1
     // ADMIN ONLY
+
     @DeleteMapping("/{id}")
     @PreAuthorize("hasRole('ADMIN')")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteCategory(@PathVariable int id)
     {
-        categoryDao.delete(id); // delete the category by id
+        Category category = categoryDao.getById(id);
+        if (category == null)
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        categoryDao.delete(id);
     }
 }
